backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
python-jose==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.7.1
httpx==0.28.1
pyotp==2.9.0
cryptography==46.0.6
orjson==3.10.14
websockets==14.2
